+++ app/src/main/java/com/clloret/speakingpractice/exercise/list/ExerciseListFragment.kt
@@ -77,7 +77,7 @@ class ExerciseListFragment : BaseFragment(), ExerciseListAdapter.ExerciseListLis
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         val binding: ExerciseListFragmentBinding = DataBindingUtil.inflate(
             inflater,
             R.layout.exercise_list_fragment, container, false
@@ -195,9 +195,9 @@ class ExerciseListFragment : BaseFragment(), ExerciseListAdapter.ExerciseListLis
         override fun onPrepareActionMode(mode: ActionMode, menu: Menu): Boolean {
             Timber.d("onPrepareActionMode")
 
-            selectionTracker?.selection?.size()?.let {
+            selectionTracker?.selection?.size()?.let { count ->
                 val menuItem = menu.findItem(R.id.action_edit)
-                menuItem.isVisible = it >= 1
+                menuItem.isVisible = count == 1
 
                 return true
             }